Address review comments - shared component, suspense
